+++ Gruntfile.js
@@ -3,7 +3,7 @@
  * Released under the MIT license.
  * www.laxarjs.org
  */
-/*global module,__dirname,__filename */
+/*global module */
 module.exports = function( grunt ) {
    'use strict';
 
@@ -13,174 +13,15 @@ module.exports = function( grunt ) {
 
    grunt.initConfig( {
       pkg: grunt.file.readJSON( 'package.json' ),
-      connect: {
+      'laxar-configure': {
          options: {
-            port: serverPort
-         },
-         default: {},
-         test: {
-            options: {
-               port: testPort,
-               keepalive: false
-            }
-         },
-         keepalive: {
-            options: {
-               port: serverPort,
-               keepalive: true
-            }
-         }
-      },
-      jshint: {
-         options: {
-            jshintrc: __dirname + '/.jshintrc'
-         }
-      },
-      karma: {
-         options: {
-            reporters: [ 'progress', 'junit' ],
-            proxies: {
-               '/base': 'http://localhost:' + testPort
-            }
-         }
-      },
-      compress: {
-         default: {
-            options: {
-               archive: '<%= pkg.name %>-<%= pkg.version %>.zip',
-               mode: 'zip'
-            },
-            files: [ {
-               src: [
-                  '*.{html,js,json}',
-                  'application/**',
-                  'bower_components/**',
-                  'includes/{controls,lib,themes,widgets}/**',
-                  'var/**',
-                  '!includes/**/{bower_components,node_modules}/**'
-               ],
-               filter: 'isFile'
-            } ]
-         }
-      },
-      laxar_application_dependencies: {
-         default: {
-            options: {},
-            dest: 'var/static/laxar_application_dependencies.js',
-            src: [ 'application/flow/*.json' ]
-         }
-      },
-      css_merger: {
-         default: {
-            src: [ 'application/flow/*.json' ]
-         }
-      },
-      cssmin: {
-         default: {
-            options: {
-               keepSpecialComments: 0
-            },
-            files: [ {
-               expand: true,
-               src: 'var/static/css/*.theme.css',
-               ext: '.theme.css'
-            } ]
-         }
-      },
-      directory_tree: {
-         application: {
-            dest: 'var/listing/application_resources.json',
-            src: [
-               'application/{flow,pages}/**/*.json',
-               'application/layouts/**/*.{css,html}'
+            flows: [
+               { target: 'main', src: 'application/flow/flow.json' }
             ],
-            options: {
-               embedContents: [
-                  'application/{flow,pages}/**/*.json',
-                  'application/layouts/**/*.html'
-               ]
-            }
-         },
-         bower_components: {
-            dest: 'var/listing/bower_components_resources.json',
-            src: [
-               'bower_components/laxar-uikit/dist/themes/default.theme/css/theme.css',
-               'bower_components/laxar-uikit/dist/controls/*/{control.json,*.theme/css/*.css}',
-               'bower_components/laxar-*-control/{control.json,*.theme/css/*.css}'
-            ],
-            options: {
-               embedContents: [
-                  'bower_components/laxar-uikit/dist/controls/*/control.json'
-               ]
-            }
-         },
-         includes: {
-            dest: 'var/listing/includes_resources.json',
-            src: [
-               'includes/themes/*.theme/{css,controls,widgets,layouts}/**/*.{css,html}',
-               'includes/controls/*/*/{control.json,*.theme/css/*.css}',
-               'includes/widgets/*/*/{widget.json,*.theme/css/*.css,*.theme/*.html}'
-            ],
-            options: {
-               embedContents: [
-                  'includes/controls/**/control.json',
-                  'includes/widgets/*/*/{widget.json,*.theme/*.html}'
-               ]
-            }
-         }
-      },
-      concat: {
-         build: {
-            src: [
-               'require_config.js',
-               'bower_components/requirejs/require.js'
-            ],
-            dest: 'var/build/require_configured.js'
-         }
-      },
-      requirejs: {
-         default: {
-            options: {
-               mainConfigFile: 'require_config.js',
-               deps: [ '../var/build/require_configured' ],
-               name: '../init',
-               out: 'var/build/bundle.js',
-               optimize: 'uglify2'
-            }
-         }
-      },
-      watch: {
-         options: {
-            livereload: liveReloadPort,
-            reload: true
-         },
-         Gruntfile: {
-            files: __filename
-         },
-         application: {
-            files: [
-               'application/**/!(scss)/*.*'
-            ]
-         },
-         libraries: {
-            files: [
-               'includes/lib/*/!(bower_components|node_modules)/**',
-               'includes/themes/*.theme/!(bower_components|node_modules)/**',
-               '!includes/**/test-results.xml'
-            ]
-         },
-         dependencies: {
-            files: [
-               '<%= directory_tree.application.src %>',
-               '<%= directory_tree.includes.src %>'
-            ],
-            tasks: [
-               'directory_tree:application',
-               'directory_tree:includes',
-               'laxar_application_dependencies'
-            ],
-            options: {
-               event: [ 'added', 'deleted' ]
+            ports: {
+               develop: serverPort,
+               test: testPort,
+               livereload: liveReloadPort
             }
          }
       }
@@ -188,48 +29,17 @@ module.exports = function( grunt ) {
 
    ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-   var path = require( 'path' );
-
-   grunt.file.expand( 'includes/widgets/*/*/widget.json' )
-      .map( path.dirname )
-      .forEach( configureArtifact.bind( null, 'widget' ) );
-
-   grunt.file.expand( 'includes/controls/**/control.json' )
-      .map( path.dirname )
-      .forEach( configureArtifact.bind( null, 'control' ) );
-
-   function configureArtifact( type, artifact ) {
-      var config = grunt.config( type + '.' + artifact ) || {};
-      config.karma = config.karma || { junitReporter: { outputFile: artifact + '/test-results.xml' } };
-      grunt.config( type + '.' + artifact, config );
-
-      grunt.config( 'watch.' + type + '_' + artifact, {
-         files: [
-            artifact + '/!(bower_components|node_modules)',
-            artifact + '/!(bower_components|node_modules)/**',
-            '!' + artifact + '/test-results.xml'
-         ]
-      } );
-   }
-
-   ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
    grunt.loadNpmTasks( 'grunt-laxar' );
-   grunt.loadNpmTasks( 'grunt-contrib-cssmin' );
-   grunt.loadNpmTasks( 'grunt-contrib-compass' );
-   grunt.loadNpmTasks( 'grunt-contrib-compress' );
-   grunt.loadNpmTasks( 'grunt-contrib-concat' );
-   grunt.loadNpmTasks( 'grunt-contrib-watch' );
 
-   ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+   // basic aliases
+   grunt.registerTask( 'test', [ 'laxar-test' ] );
+   grunt.registerTask( 'build', [ 'laxar-build' ] );
+   grunt.registerTask( 'develop', [ 'laxar-develop' ] );
 
-   grunt.registerTask( 'server', [ 'connect:default' ] );
-   grunt.registerTask( 'build', [ 'directory_tree', 'laxar_application_dependencies' ] );
-   grunt.registerTask( 'optimize', [ 'build', 'css_merger', 'cssmin', 'concat', 'requirejs' ] );
-   grunt.registerTask( 'test', [ 'connect:test', 'widgets' ] );
-   grunt.registerTask( 'dist', [ 'optimize', 'compress' ] );
-   grunt.registerTask( 'start', [ 'build', 'server', 'watch' ] );
-   grunt.registerTask( 'start-no-watch', [ 'build', 'connect:keepalive' ] );
+   // additional (possibly) more intuitive aliases
+   grunt.registerTask( 'optimize', [ 'laxar-dist' ] );
+   grunt.registerTask( 'start', [ 'laxar-develop' ] );
+   grunt.registerTask( 'start-no-watch', [ 'laxar-develop-no-watch' ] );
 
    grunt.registerTask( 'default', [ 'build', 'test' ] );