+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#22](https://github.com/LaxarJS/shop-demo/issues/22): project: switched to flow-based grunt tasks
 - [#20](https://github.com/LaxarJS/shop-demo/issues/20): added Travis-CI build integration
     + NEW FEATURE: see ticket for details